Изм 1.9.11 Рефакторинг UI and Data Выполнена проверка и слияние изменений в репозитории и VW Убраны лишние теперь и временные заглушки Обновлены данные заглушки для списка последних действий по книгам (эта заглушка использвается для списка в главном фрагменте и для списка во фрагменте общего списка событий Выявлена небольшая проблемка с генератором заглушки в части насыщения данными по дням дальше даты окончания представления. Вплане решить в следующем сводном PR (если к тому времени не будет готов ответ сервера)
