+++ app/src/main/java/ru/vdv/myapp/myreadersdiary/ui/statistics/SummaryStatisticsViewModel.kt
@@ -4,6 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import ru.vdv.myapp.myreadersdiary.domain.*
 import ru.vdv.myapp.myreadersdiary.ui.common.BaseViewModel
+import java.text.SimpleDateFormat
+import java.util.*
 
 class SummaryStatisticsViewModel : BaseViewModel() {
 
@@ -20,11 +22,17 @@ class SummaryStatisticsViewModel : BaseViewModel() {
 
     //events
     private val _prepareEventList = MutableLiveData<List<WeekEvent>>().apply {
-        repository.getSummaryEventData(object : CallBack<List<WeekEvent>> {
+        val startData = Date(convertDateToLong("2020.11.13 10:25"))
+        repository.getRandomSummaryEventData(startData, object : CallBack<List<WeekEvent>> {
             override fun onResult(result: List<WeekEvent>) {
                 value = result
             }
         })
+    }
+
+    private fun convertDateToLong(date: String): Long {
+        val df = SimpleDateFormat("yyyy.MM.dd HH:mm")
+        return df.parse(date).time
     }
     val prepareEventList: LiveData<List<WeekEvent>> = _prepareEventList
 }
